Update and Fixed player movement
Diagonals now work! Speed stays consistent when moving in multiple
directions. The issue was that the checks for which keys were down were
triggered when it checked for both keys. So, with LEFT and UP both down,
during the LEFT check player moved LEFT **and** UP, and then it moved
again on the check for UP. 

Now just need to balance the speed. It seems a little slow at the
moment.